requirements.txt
alembic==1.8.1
anyio==3.6.2
asyncpg==0.26.0
click==8.1.3
databases==0.6.1
fastapi==0.79.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
Mako==1.2.3
MarkupSafe==2.1.1
orjson==3.8.1
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
starlette==0.19.1
typing_extensions==4.4.0
uvicorn==0.18.2
uvloop==0.14.0
